Update User Components and File Structure
