+++ sprost/src/scripts/Authentication.tsx
