+++ sprost/src/scripts/app/components/Paragraph.tsx
@@ -1,0 +1,10 @@
+import React, { FC } from "react";
+import { Paragraph as ParagraphType } from "../../../types/components/Paragraph";
+
+const Paragraph: FC<{ component: ParagraphType }> = ({ component }) => {
+	return <p>
+		{component.message}
+	</p>;
+};
+
+export default Paragraph;
