+++ sprost/src/scripts/sprost/editors/Header.tsx
