+++ sprost/src/scripts/sprost/modals/NewComponent.tsx
