+++ sprost/src/scripts/sprost/modals/NewView.tsx
